+++ khorosjx/places/__init__.py
@@ -1,0 +1,17 @@
+# -*- coding: utf-8 -*-
+"""
+:Package:        khorosjx.places
+:Synopsis:       This package includes custom exceptions and accompanying function
+:Usage:          ``import khorosjx.places``
+:Example:        TBD
+:Created By:     Jeff Shurtliff
+:Last Modified:  Jeff Shurtliff
+:Modified Date:  22 Dec 2019
+"""
+# Define all modules that will be imported with the "import *" method
+__all__ = ['base', 'blogs', 'spaces']
+
+# Always import the primary core module, the places base modules and also the exceptions and handlers modules
+from .. import core
+from ..errors import exceptions, handlers
+from . import base as places_core